split out schedules.py
