+++ corehq/apps/app_manager/models/__init__.py
@@ -47,7 +47,6 @@ from .common import (
     FormBase,
     FormDatum,
     FormLink,
-    FormSchedule,
     FormSource,
     GlobalAppConfig,
     GraphAnnotations,
@@ -78,9 +77,6 @@ from .common import (
     ReportAppFilter,
     ReportModule,
     SavedAppBuild,
-    SchedulePhase,
-    SchedulePhaseForm,
-    ScheduleVisit,
     ShadowForm,
     ShadowModule,
     SortElement,
